+++ src/Loader.php
@@ -8,7 +8,7 @@ class Loader
 {
     public function __construct()
     {
-        (new ConfigFile())->createConfigFileIfDoesNotExists(getcwd() . '/wp-config-local.php');
+        //((new \SourceBroker\DeployerExtendedWordpress\Drivers\Driver()))->createConfigFileIfDoesNotExists(getcwd() . '/config');
         /** @noinspection PhpIncludeInspection */
         require_once 'recipe/common.php';
         new Load([