+++ src/Drivers/EnvDriver.php
@@ -27,9 +27,9 @@ class EnvDriver
     public function getInstanceName($absolutePathWithConfig = null): string
     {
         $this->loadEnv($absolutePathWithConfig);
-        $instanceName = strtolower(getenv('WP_ENV'));
+        $instanceName = strtolower(getenv('WP_INSTANCE') ? getenv('WP_INSTANCE') : getenv('WP_ENV'));
         if (empty($instanceName)) {
-            throw new \Exception("\nWP_ENV env variable is not set. \nIf this is your local instance then please put following line: \nWP_ENV=development \nin configuration file: ' . $absolutePathWithConfig . '\n\n");
+            throw new \Exception("\nWP_INSTANCE/WP_ENV env variable is not set. \nIf this is your local instance then please put following line: \nWP_ENV=development (or WP_INSTANCE=dev if you have instance based settings)\nin configuration file: ' . $absolutePathWithConfig . '\n\n");
         }
         return $instanceName;
     }
@@ -43,7 +43,7 @@ class EnvDriver
         $absolutePathWithConfig = rtrim($absolutePathWithConfig, DIRECTORY_SEPARATOR);
         $envFilePath = $absolutePathWithConfig . '/.env';
         if (file_exists($envFilePath)) {
-            (new Dotenv(true))->load($envFilePath);
+            (new Dotenv(true))->loadEnv($envFilePath, 'WP_INSTANCE');
         } else {
             throw new \Exception('Missing file "' . $envFilePath . '".');
         }