+++ CHANGELOG.rst
@@ -7,7 +7,8 @@ master
 
 a) [BUGFIX][BREAKING] Rename task names because Windows compatibility.
 b) [TASK][BREAKING] Add compatibility with new deployer-instance.
-
+c) [TASK] Do not allow to pull, push, copy, link media and database to live instance.
+d) [TASK] Add "export $(cat .env | xargs) " to db:import:post_command:wp_domains to have possibility to set php version.
 
 1.0.0
 ~~~~~