+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ a) [BUGFIX][BREAKING] Rename task names because Windows compatibility.
 b) [TASK][BREAKING] Add compatibility with new deployer-instance.
 c) [TASK] Do not allow to pull, push, copy, link media and database to live instance.
 d) [TASK] Add "export $(cat .env | xargs) " to db:import:post_command:wp_domains to have possibility to set php version.
+e) [TASK] Update deployer-extended-media, deployer-extended-database, deployer-instance, deployer-extended.
 
 1.0.0
 ~~~~~